VM->ENCLAVE FFI for handle_type

+++ go-cosmwasm/api/lib.go
@@ -209,7 +209,7 @@ func Handle(
 	runtime.LockOSThread()
 	defer runtime.UnlockOSThread()
 
-	res, err := C.handle(cache.ptr, id, p, m, db, a, q, u64(gasLimit), &gasUsed, &errmsg, s, handleType)
+	res, err := C.handle(cache.ptr, id, p, m, db, a, q, u64(gasLimit), &gasUsed, &errmsg, s, uint8(handleType))
 	if err != nil && err.(syscall.Errno) != C.ErrnoValue_Success {
 		// Depending on the nature of the error, `gasUsed` will either have a meaningful value, or just 0.
 		return nil, uint64(gasUsed), errorWithMessage(err, errmsg)
